+++ src/app/shared/shared.module.ts
@@ -1,12 +1,15 @@
+// Libs
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-
 import { ReactiveFormsModule } from '@angular/forms';
+
+// Angular Material Modules
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
+// Components
 import { ButtonComponent } from './components/button/button.component';
 import { InputComponent } from './components/input/input.component';
